grounded-pest-app/requirements.txt
streamlit
requests
python-dotenv
openai>=1.43.0
azure-storage-blob
